# vms-backend/openapi.yaml
openapi: 3.0.3
info:
  title: VMS API
  version: 0.1.0
  description: Visitor Management System — event badge scanning on a closed LAN.
paths:
...
  /api/v1/badges/card:
    post:
      operationId: badges_card_create
      description: |-
        Takes the raw badge token returned once by `POST /visitors` and returns a CR80 card as a PDF. The token draws the QR and is then discarded; it is never stored, and the visitor is resolved from its digest.

        There is no GET equivalent. Only the digest is kept, so the server cannot reprint a card without being handed the token again.
      summary: Render a single badge PDF
      tags:
      - badges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgeCardRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BadgeCardRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BadgeCardRequestRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: A CR80 badge PDF.
        '404':
          description: No visitor holds that token.
  /api/v1/badges/export:
    post:
      operationId: badges_export_create
      description: |-
        NON-DESTRUCTIVE. Every QR in the file is the code ALREADY on that visitor's card, recomputed from the derived token, so nothing is reissued and no printed card stops scanning. Exporting twice produces an identical file.

        A printing worklist: a title banner naming the event, then No. / Name / Country / Organization / Category / Badge Serial / Photo / QR Code / Registered / QR payload from row 5, with the banner frozen. The payload is the exact string the QR encodes, so a card producer can re-render it at their own size.

        It therefore contains working credentials, and a second sheet says so, because a spreadsheet outlives the click that made it.

        Omit `visitor_ids` to take every visitor who has not been deleted.
      summary: Export badges as .xlsx with photos and QR codes
      tags:
      - badges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgeExportRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BadgeExportRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BadgeExportRequestRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: An .xlsx workbook with one QR per visitor.
  /api/v1/badges/reissue:
    post:
      operationId: badges_reissue_create
      description: |-
        DESTRUCTIVE. Every listed visitor is given a NEW badge token, which invalidates the QR on any card already printed for them.

        Returns the raw tokens so a client can render the QR itself — on screen, or into its own layout. The server cannot return the token of an existing card, only of one it has just minted.

        This response is the ONLY copy. Nothing stores it.
      summary: Reissue badges and return their raw tokens
      tags:
      - badges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReissueResponse'
          description: ''
  /api/v1/badges/reissue-sheet:
    post:
      operationId: badges_reissue_sheet_create
      description: |-
        DESTRUCTIVE. Every listed visitor is given a NEW badge token, which invalidates the QR on any card already printed for them, including one they are currently wearing.

        This is not a choice the endpoint makes: the raw token exists only at the moment it is created, so reprinting is impossible and reissuing is the only thing the server can do.

        Nine cards per A4 sheet with cut marks; more than nine paginates. A run of exactly one comes back as a single 54x85.6mm card page instead, since one card on A4 wastes the sheet.
      summary: Reissue badges and render a printable PDF
      tags:
      - badges
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BadgeReissueSheetRequestRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: An A4 sheet of badges as a PDF, or a single CR80 card page
            when exactly one visitor was requested.
  /api/v1/badges/roster.xlsx:
    get:
      operationId: badges_roster_export_retrieve
      description: |-
        Who is registered, as a printing worklist: a title banner naming the event, then No. / Name / Country / Organization / Photo / QR Code / Registered from row 5, with the banner frozen.

        NON-DESTRUCTIVE. Nothing is reissued and no card stops working. Each QR is the visitor's existing badge token, derived rather than stored, so the code drawn here is the code already on their card.

        It therefore contains working credentials. Use POST /badges/export instead only when an outside card producer needs freshly minted tokens and every current card is being replaced.
      summary: Export the visitor roster as .xlsx
      tags:
      - badges
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: An .xlsx workbook.
...
  /api/v1/visitors:
    get:
      operationId: visitors_list
      summary: List visitors
      parameters:
      - in: query
        name: category
        schema:
          type: string
          enum:
          - normal
          - vip
        description: Filter by badge category.
      - in: query
        name: country
        schema:
          type: string
        description: Exact country match, case-insensitive.
      - in: query
        name: is_active
        schema:
          type: boolean
        description: Filter by badge state. `false` lists deactivated visitors.
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: with_tokens
        schema:
          type: boolean
        description: Include `badge_token` on every row. Off by default. Each token
          is a working credential, so this turns one request into a set of usable
          badges -- ask for it only where the codes are the point, such as a print
          queue that draws the real QR. The same codes are already in the .xlsx export.
      tags:
      - visitors
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitorList'
          description: ''
    post:
      operationId: visitors_create
      description: Returns `badge_token`, the raw value for the QR code. It is stored
        only as a SHA-256 digest, so this response is the single chance to capture
        it — generate the badge PDF from it immediately.
      summary: Register a visitor and issue a badge
      tags:
      - visitors
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitorIssuedRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitorIssuedRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/VisitorIssuedRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorIssued'
          description: ''
  /api/v1/visitors/{id}:
    get:
      operationId: visitors_retrieve
      summary: Visitor detail, with scan history
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this visitor.
        required: true
      tags:
      - visitors
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorDetail'
          description: ''
    patch:
      operationId: visitors_partial_update
      summary: Edit a visitor
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this visitor.
        required: true
      tags:
      - visitors
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedVisitorRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedVisitorRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedVisitorRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visitor'
          description: ''
    delete:
      operationId: visitors_destroy
      summary: Soft-delete a visitor (also deactivates the badge)
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this visitor.
        required: true
      tags:
      - visitors
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    BadgeCardRequestRequest:
      type: object
      description: |-
        What the registration receipt posts back to get a printable card.

        The raw token, and only the raw token. The server resolves the visitor from
        its SHA-256 digest exactly as `/scans` does, so this is not an oracle: a
        caller who does not already hold the token gets nothing, and one who does is
        holding the badge anyway.
      properties:
        token:
          type: string
          minLength: 1
          description: The raw badge token, as returned once by POST /visitors. Not
            stored; used to draw the QR and then discarded.
          maxLength: 128
      required:
      - token
    BadgeExportRequestRequest:
      type: object
      description: |-
        Which visitors go into the credential spreadsheet.

        Every one of them gets a NEW token, for the same reason the sheet does: the
        server keeps only a digest, so a scannable QR can be minted but never
        recovered. Omit the list to take everyone who is still registered.
      properties:
        visitor_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Visitors to reissue and export, in row order. Omit to export
            every visitor who has not been deleted.
          maxItems: 250
    BadgeReissueSheetRequestRequest:
      type: object
      description: |-
        Which visitors to put on the sheet.

        Every one of them gets a NEW token — see the endpoint description. There is no
        way to reprint an existing card, because the raw token it carries was never
        stored.
      properties:
        visitor_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Visitors to reissue and print, in the order they appear on
            the sheet.
          maxItems: 250
      required:
      - visitor_ids
...
    CategoryEnum:
      enum:
      - normal
      - vip
      type: string
      description: |-
        * `normal` - Normal
        * `vip` - VIP
...
    ReissueResponse:
      type: object
      properties:
        issued:
          type: array
          items:
            $ref: '#/components/schemas/ReissuedBadge'
          readOnly: true
      required:
      - issued
...
    VisitorIssuedRequest:
      type: object
      description: |-
        The response to `POST /visitors` — the one and only sight of the raw token.

        `badge_token` is what goes into the QR. It is not stored anywhere and cannot
        be recovered: the caller must print the card now (CLAUDE.md constraint #3).
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: string
          minLength: 1
          maxLength: 100
        organization:
          type: string
          maxLength: 200
        photo:
          type: string
          format: binary
        category:
          $ref: '#/components/schemas/CategoryEnum'
      required:
      - country
      - full_name
      - photo
    VisitorList:
      type: object
      description: |-
        Documents `GET /visitors`, where the token is conditional.

        SCHEMA ONLY -- nothing serialises through this. The view picks
        `VisitorSerializer` or `VisitorWithTokenSerializer` per request depending on
        `?with_tokens=`, and OpenAPI has no way to say "this response shape depends
        on that query parameter". Declaring the union honestly, with `badge_token`
        optional, is closer to the truth than either half alone: a client is told the
        field may be absent and has to check, which is exactly the situation.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        full_name:
          type: string
          maxLength: 200
        country:
          type: string
          maxLength: 100
        organization:
          type: string
          maxLength: 200
        photo:
          type: string
          format: uri
        category:
          $ref: '#/components/schemas/CategoryEnum'
        badge_serial:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        badge_token:
          type: string
          description: Present only when the request asked for `with_tokens=true`.
            The raw badge token for the QR code, stable for the life of the badge.
      required:
      - badge_serial
      - country
      - created_at
      - full_name
      - id
      - is_active
      - photo
      - updated_at
...
  securitySchemes:
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Permanent token from `POST /api/v1/devices/pair`, sent as `Authorization:
        Device <token>`.'
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT